+++ src/config/mdns/discover.rs
@@ -8,6 +8,6 @@ pub struct MdnsDiscoverArgs {
     #[command(flatten)]
     pub service_type: ServiceTypeArgs,
     /// How long in ms to attempt to discover services before shutdown
-    #[arg(long, default_value_t = 4000)]
+    #[arg(long, default_value_t = 1000)]
     pub timeout_ms: u64,
 }